requirements.txt
gradio==4.31.0
langchain==0.3.0
langchain-openai==0.2.0
faiss-cpu==1.8.0
openai==1.30.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.3.0
numpy==1.26.0
sentence-transformers==2.2.2